requirements.txt
Django==2.1.2
psycopg2-binary==2.7.5
pytz==2018.6
django-crispy-forms==1.7.2
Pillow==5.3.0
reportlab==3.5.10
djangorestframework==3.7.7
django-rest-swagger==2.1.2
coreapi==2.3.3
docx-mailmerge==0.4.0
django-mptt==0.10.0